osbuild: `default_main` entrypoint for modules
This implements a default entrypoint for modules which they are required
to implement. This entrypoint makes every module an executable that
takes two command line arguments: `-s/--schema` to return its schema,
and `-i/--info` to return its documentation. The `-c/--capabilities` to
return necessary capabilities. `-a/--all` returns the information of all
other options as a single JSON object. This object is the same as the
representation used in `osbuild/meta.py`.

Having this allows to decouple osbuild from its modules more strictly,
currently osbuild has to import modules to ask about this information;
this blocks any option of implementing modules in different languages
that might be better suited to some of the tasks.