osbuild: `default_main` to parse modules
Change `ModuleInfo` to execute the module with the `-a` flag to receive
the information necessary to register it.

This removes the dependency on modules needing to be Python files and
removes the `ast`-related code.

One minor disadvantage: the schema validation error in this class can
now no longer show the actual line number in the module anymore, instead
it shows the linenumber in the schema again.